+++ wire-ios-request-strategy/Sources/Payloads/Conversation/Payload+Conversation.swift
@@ -191,29 +191,6 @@ extension Payload {
 
     // MARK: - Events
 
-    struct UpdateConversationMLSWelcome: Codable {
-
-        enum CodingKeys: String, CodingKey {
-            case id = "conversation"
-            case qualifiedID = "qualified_conversation"
-            case from
-            case qualifiedFrom = "qualified_from"
-            case timestamp = "time"
-            case type
-            case data
-        }
-
-        // This is currently the id of the self conversation.
-        let id: UUID
-
-        let qualifiedID: QualifiedID?
-        let from: UUID
-        let qualifiedFrom: QualifiedID?
-        let timestamp: Date
-        let type: String
-        let data: String
-    }
-
     struct UpdateConversationProtocolChange: CodableEventData {
         static var eventType: ZMUpdateEventType {
             return .conversationConnectRequest
